# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Scribe API – Enterprise Blogging & Content Management Engine by Shivani Dubey","description":"### Production-Ready Blogging Engine Backend\nAn enterprise-grade, high-performance RESTful content management API built using Spring Boot 3, Java 23, and Spring Security.\n\n#### Core Architectural Implementations Highlighted:\n* **Stateless JWT Security Filter Matrix**: Fine-grained role guards (ROLE_ADMIN / ROLE_NORMAL) keeping internal memory overhead flat.\n* **N+1 Database Query Mitigation**: Custom repository layers utilizing JOIN FETCH operations to fetch relational dependencies in single transaction steps.\n* **Isolated Filesystem Media Tier**: Local disk path protection structures parsing image uploads securely away from the source code via randomized UUID identifiers.\n* **Global Exception Envelopes**: @RestControllerAdvice interceptors formatting data validations smoothly into structured JSON tracking models.\n\n--- \n* **Developer Profile**: Shivani Dubey | [LinkedIn](www.linkedin.com/in/shivanis5) | [Email](mailto:work.shivanidubey@gmail.com)\n\n*To unlock and execute secure endpoints, click the Authorize button on the right and input a valid JWT Bearer token.*","license":{"name":"MIT Open Source License","url":"https://opensource.org"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local Sandbox (Development)"}],"security":[{"Bearer Authentication":[]}],"paths":{"/api/v1/users/{userId}":{"get":{"tags":["user-controller"],"operationId":"getSingleUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/posts/{postId}":{"get":{"tags":["post-controller"],"operationId":"getPostById","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostDto"}}}}}},"put":{"tags":["post-controller"],"operationId":"updatePost","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostDto"}}}}}},"delete":{"tags":["post-controller"],"operationId":"deletePost","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/categories/{catId}":{"get":{"tags":["category-controller"],"operationId":"getCategorybyId","parameters":[{"name":"catId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}}},"put":{"tags":["category-controller"],"operationId":"updateCategory","parameters":[{"name":"catId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}}},"delete":{"tags":["category-controller"],"operationId":"deleteCategory","parameters":[{"name":"catId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id"}},{"name":"sortDir","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/posts/{postId}/comments":{"post":{"tags":["comment-controller"],"operationId":"createComment","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentDto"}}}}}}},"/api/v1/posts/user/{userId}/category/{categoryId}":{"post":{"tags":["post-controller"],"operationId":"createPost","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostDto"}}}}}}},"/api/v1/posts/image/upload/{postId}":{"post":{"tags":["post-controller"],"operationId":"uploadPostImage","parameters":[{"name":"postId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"image":{"type":"string","format":"binary"}},"required":["image"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostDto"}}}}}}},"/api/v1/categories":{"get":{"tags":["category-controller"],"operationId":"getAllCategories","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"catId"}},{"name":"sortDir","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}}}},"post":{"tags":["category-controller"],"operationId":"createCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"createToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtAuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthResponse"}}}}}}},"/api/v1/posts":{"get":{"tags":["post-controller"],"operationId":"getAllPost","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"postId"}},{"name":"sortDir","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}}}}},"/api/v1/posts/user/{userId}":{"get":{"tags":["post-controller"],"operationId":"getPostsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostDto"}}}}}}}},"/api/v1/posts/search/{keywords}":{"get":{"tags":["post-controller"],"operationId":"searchPostByTitle","parameters":[{"name":"keywords","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostDto"}}}}}}}},"/api/v1/posts/image/{imageName}":{"get":{"tags":["post-controller"],"operationId":"downloadImage","parameters":[{"name":"imageName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/posts/category/{categoryId}":{"get":{"tags":["post-controller"],"operationId":"getPostsByCategory","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostDto"}}}}}}}},"/api/v1/auth/current-user":{"get":{"tags":["auth-controller"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/comments/{commentId}":{"delete":{"tags":["comment-controller"],"operationId":"deleteComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}}},"components":{"schemas":{"RoleDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":50,"minLength":4},"email":{"type":"string"},"password":{"type":"string","maxLength":255,"minLength":6,"writeOnly":true},"about":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/RoleDto"},"uniqueItems":true}},"required":["about","email","name","password"]},"CategoryDto":{"type":"object","properties":{"categoryId":{"type":"integer","format":"int32"},"categoryTitle":{"type":"string","maxLength":2147483647,"minLength":4},"categoryDescription":{"type":"string","maxLength":2147483647,"minLength":10}},"required":["categoryDescription","categoryTitle"]},"CommentDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"content":{"type":"string","maxLength":1000,"minLength":0},"user":{"$ref":"#/components/schemas/UserDto"}},"required":["content"]},"PostDto":{"type":"object","properties":{"postId":{"type":"integer","format":"int32"},"title":{"type":"string","maxLength":100,"minLength":4},"content":{"type":"string","maxLength":10000,"minLength":10},"imageName":{"type":"string"},"addedDate":{"type":"string","format":"date-time"},"category":{"$ref":"#/components/schemas/CategoryDto"},"user":{"$ref":"#/components/schemas/UserDto"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentDto"},"uniqueItems":true}},"required":["content","title"]},"JwtAuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["password","username"]},"JwtAuthResponse":{"type":"object","properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"}}},"UserResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"lastPage":{"type":"boolean"}}},"PostResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/PostDto"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"lastPage":{"type":"boolean"}}},"CategoryResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"lastPage":{"type":"boolean"}}},"ApiResponse":{"type":"object","properties":{"message":{"type":"string"},"success":{"type":"boolean"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","description":"Provide a valid JSON Web Token (JWT) to authorize requests against secure paths.","name":"Bearer Authentication","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}